refactore:  db.ts->db.config.ts
