+++ backend/src/middlewares/permissions/team-permission.middleware.ts
@@ -1,5 +1,5 @@
 import { Request, Response, NextFunction } from "express";
-import {prisma} from "../../config/db";
+import {prisma} from "../../config/db.config";
 import { StatusCodes } from "http-status-codes";
 import { TEAMS_CONSTANTS } from "../../constants/teams.constants";
 import validateUser from "../validation/user.middleware";
